角速度をvelocityctrlじゃなくてIMUで計算するようにした

+++ Core/Inc/Encoder.hpp
@@ -13,21 +13,21 @@
 class Encoder{
 
 private:
-	float cnt_l_, cnt_r_;
-	float distance_; //[mm]
-	long total_cnt_l_, total_cnt_r_;
-	float total_distance_;
+	double cnt_l_, cnt_r_;
+	double distance_; //[mm]
+	double total_cnt_l_, total_cnt_r_;
+	double total_distance_;
 
 public:
 	Encoder();
 	void init();
 	void updateCnt();
-	void getCnt(float &, float &);
-	float getDistance();
-	float getTotalDistance();
+	void getCnt(double &, double &);
+	double getDistance();
+	double getTotalDistance();
 	void clearDistance();
 	void clearCnt();
-	long getTotalCnt();
+	double getTotalCnt();
 	void clearTotalCnt();
 };
